Number of Available R Packages and R Versions from MRAN

MRAN has an archive of Snapshots of CRAN dating back to September 17 2014. These functions return the number of available packages and the available R version according to the snapshot of https://cran.r-project.org on MRAN.

n_available_packages(date = Sys.Date())

available_r_version(date = Sys.Date())

Arguments

date

the date of the snapshot to be used. It can be a Date object or a character in the format %Y-%m-%d.

Value

the number of available packages as an integer or the R version number as a character

Details

MRAN has data starting from September 17 2014. Data for a few selected dates before September 17 2014 can be obtained from the dataset Ecdat::CRANpackages. A more complete dataset ranging from 2001 until today can be obtained with get_cran_history().

Note that for some dates there is no snapshot on MRAN. The function will return an error in those cases.

See also