Protein Consumption in European Countries

Protein Consumption from various sources in European countries in unspecified units. The exact year of data collection is not known but the oldest known publication of the data is from 1973.

protein

Format

Data frame with 25 rows and 10 variables:

country

name of the country

red_meat

red meat

white_meat

white meat

eggs

eggs

milk

milk

fish

fish

cereals

cereals

starch

starchy foods

nuts

pulses, nuts, oil-seeds

fruit_veg

fruits, vegetables

Source

The data have been downloaded from https://raw.githubusercontent.com/jgscott/STA380/master/data/protein.csv

They come from the following book:

Hand, D. J. et al. (1994). A Handbook of Small Data Sets, Chapman and Hall, London.

Chapter 360, p. 297.

In the book, it is stated that the data have first been published in

Weber, A. (1973). Agrarpolitik im Spannungsfeld der internationalen Ernährungspolitik, Institut für Agrarpolitik und Marktlehre, Kiel.